requirements-travis.txt
django-debug-toolbar==0.9.4
django-nose==1.1
docutils==0.10
nose==1.2.1
openpyxl==1.6.1
wsgiref==0.1.2
xlrd==0.9.0
chardet==2.1.1
South==0.8.1
